putfile(): disable echo in target shell before sending
A --put of a large file will result in the target continuing to echo the
commands that serio uses to copy the data to the target, even after serio
exits.  If you quickly connect minicom to the target after a large --put
completes, you will see the echoed commands.

Also convert some leading spaces to tabs.