switch delim definition back to complex form, use universally
Tim's use of a complex value for variable "delim" was meant to allow
the serio program to be transferred by serio.  This was simplified
in 338d455944e3129280d1f6f640c21f00560f7b02 since the motivation
was not documented.  Switch back to using a complex form (slightly
different from the way Tim orginally did it).

Also replace the previous value for "DELIM" in class RemoteExecute
with this complex value, and use "DELIM" in cmd() instead of having
a local "delim".