invoke remote shell to execute --cmd, use execute() for cmd()
Invoke a remote shell to execute --cmd.  The exit status of serio
is set to the exit status of the remote shell.

Add option -N, no-remote-shell to instead use Native (existing) shell.

Change cmd() to use execute() to send the command to the remote
system.  This should handle the case where the remote system
shell prompt arrives while sending the command or arrives after
sending the command.

This is a first shot at invoking a remote shell.  There may be some
rough edges or bugs to find and fix.