Base changes for multiple file copy
These changes were submitted as RFC and will be partially
completed by the following commit from Frank Rowand.

Recursive copy, Wild card copy etc has been added
Retain the permissions of the destination file as it is having in
source.  If source is a file and destination is a folder, a file
with the name of source will be created in destination with the
same name.

Signed-off-by: Vishnu Sankar vishnuocv@gmail.com