Read prompt after each transmitted line, in putfile
The prevents data overruns when doing putfile on
my setup, with io_time=0.

This is faster (and likely more robust) than
using io_time as a delay, but still prevents
overruns on the target, when doing a putfile.

Performance to put a 10000-byte file is:
 -P (io_time 0.1) => 41 seconds
 -t 0.05 (io_time 0.05) => 21 seconds
waiting for prompt, but using io_time=0 => 7 seconds

With io_time=0 and no wait for the prompt, serio hangs
with the target stuck at the secondary shell prompt
on my setup (beaglebone black, connected via Sony
debug board).  The target misses some of the data,
and since it is in the middle of an echo of quoted
material, the second quote is missed.

Signed-off-by: Tim Bird <tim.bird@sony.com>