Add time_print routine for debugging serio operations
This prints a time-stamped line for most of the interactions
with the target.  This is handy to see what gets sent to the
target, and when it is sent.  It's very verbose, and can only
be turned on by changing the definition of "debug" in the source.

To turn it on, change the line to "debug = 1".

Signed-off-by: Tim Bird <tim.bird@am.sony.com>