serio: clean up usage() messages, NOTE: 5f772ad1 was mislabeled
Some usage() message cleanups that were missed in 5f772ad1.

*****  NOTICE  *****

A big set of changes inadvertently were collapsed into commit:

   5f772ad1f15f serio: revert prompt consumption from b958843ceb1f

The serio command list was changed to be more like ssh and scp.

You need to create two soft links:

   ln -s serio sercp
   ln -s serio sersh

sercp is analagous to scp

sersh is analagous to ssh

The new syntax is shown in the usage messages:

   $sercp -h

      Usage: sercp [OPTIONS] [[user1@]host1:]file1 ... [[user2@]host2:]file2

   $ sersh -h

      Usage: sersh [OPTIONS] [user @]hostname command

This is an initial implementation and there are likely to be some
corner cases where serio does not behave exactly like scp and ssh.

Signed-off-by: Frank Rowand <frowand.list@gmail.com>