serio: recursive copy needs leading path truncated at destination
Recursive copy should copy directory tree rooted at the source directory,
not the entire path of the source directory.  For example, with source
of 'from/a/', the root of the directory tree created at the destination
should be 'a/' instead of 'from/a/'.

$ find from/a
from/a
from/a/A
from/a/b
from/a/b/C
from/a/b/c
from/a/b/B
from/a/B

Incorrect result before fixing:

  $ sercp -v -r from/a ttyUSB0:/home/serio-test
  Uploaded 2 bytes from from/a/A to /home/serio-test/from/a/A
  Uploaded 0 bytes from from/a/B to /home/serio-test/from/a/B
  Uploaded 2 bytes from from/a/b/C to /home/serio-test/from/a/b/C
  Uploaded 0 bytes from from/a/b/B to /home/serio-test/from/a/b/B

Correct result after fixing:

  $ sercp -v -r from/a ttyUSB0:/home/serio-test
  Uploaded 2 bytes from from/a/A to /home/serio-test/a/A
  Uploaded 0 bytes from from/a/B to /home/serio-test/a/B
  Uploaded 2 bytes from from/a/b/C to /home/serio-test/a/b/C
  Uploaded 0 bytes from from/a/b/B to /home/serio-test/a/b/B

Reported-by: Vishnu Sankar vishnuocv@gmail.com
Signed-off-by: Frank Rowand <frowand.list@gmail.com>