log comment fix: 5f772ad1f15f -- convert serio, serio-test.sh to ssh/scp syntax
The log for commit 5f772ad1f15f says:

   serio: revert prompt consumption from b958843ceb1f

   Patch fragments 2 and 3 are indeed that change (the fragments have
   comments talking about "consume prompt").

The rest of the changes in this commit where meant to be a separate
commit.  Those changes are adding the syntax used by ssh and scp
to specify host name and file names.

serio-test.sh is also changed to match the serio changes.

Typical serio-test usages:

   serio-test.sh -R /home/serio-test ttyACM0
   serio-test.sh -R /home/serio-test ttyUSB0

The big changes hidden in this commit are changing serio from the
old command line syntax to syntax more like ssh and scp.

You need to create two soft links:

   ln -s serio sercp
   ln -s serio sersh

(I am open to suggestions for different names.)

sercp is analagous to scp

sersh is analagous to ssh

The new syntax is shown in the usage messages:

   $sercp -h

   Usage: sercp [OPTIONS] [[user1@]host1:]file1 ... [[user2@]host2:]file2

   $ sersh -h

   Usage: sersh [OPTIONS] [user @]hostname command

You can use ttyUSB0 and ttyACM0 for host1, host2, and hostname.

You can _not_ use /dev/ttyACM0 for host1 and host2 because the '/'
before a ':' is how scp determines that you are trying to copy
a file with a ':' in the name instead of specifying a host name.

You can use either ttyUSB0 or /dev/ttyUSB0 for hostname.

This is an initial implementation and there are likely to be some
corner cases where serio does not behave exactly like scp and ssh.

serio-test.sh passes for me on:

   - dragonboard with bash
   - beaglebone black debian
   - beaglebone black angstrom

serio-test.sh has an issue for:

   - dragonboard with busybox sh
       - the clean up attempt to rm files on the remote does
         not function properly

target board software:

   dragonboard, Linux 4.10.0-rc3-dirty

   beaglebone, debian 8, Linux 4.4.36-ti-r72, uuencode is busybox

   beaglebone, angstrom v2012.12, Linux 3.8.13

Dragonboard is via sony debug board /dev/ttyUSB0, breakout wires to rs-232

beaglebone blacks are via /dev/ttyACMx, usb cable to beaglebone