sercp: Add support for full device paths
Add a new argument (-d, --device) to support a full device path.
Due to file argument parsing constraints, sercp cannot handle
certain serial paths used as the host for the source or destination.

This feature makes it possible to use a full path, with slashes and
colons, to specify the serial port for the operation.  Use something
like:
 $ sercp -d /dev/serial/by-path/pci-0000:08:00-usb-0:3:1.4 file1 serial:/tmp

The reason for this is that in some cases, using the full path is more
convenient than the single-word device.  For example, the full path can
uniquely identify a serial port on a USB port, that might change device
names as USB ports are connected and disconnected.

Signed-off-by: Tim Bird <tim.bird@sony.com>